# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Golang template project",
        "title": "Golang Template API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/api",
    "paths": {
        "/greet": {
            "get": {
                "description": "Returns a greeting message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "greet"
                ],
                "summary": "Greet endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "default": "World",
                        "description": "Name to greet",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GreetResponse"
                        }
                    }
                }
            }
        },
        "/rss/spiegel/export": {
            "get": {
                "description": "Exports RSS headlines in CSV or JSON format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/csv"
                ],
                "tags": [
                    "rss"
                ],
                "summary": "Export SPIEGEL RSS headlines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Export format (json or csv)",
                        "name": "format",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter headlines by keyword",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of headlines to export",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rss/spiegel/latest": {
            "get": {
                "description": "Fetches the most recent headline from SPIEGEL RSS feed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rss"
                ],
                "summary": "Get latest SPIEGEL RSS headline",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shared.RssHeadline"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rss/spiegel/top5": {
            "get": {
                "description": "Fetches the top N headlines from SPIEGEL RSS feed (max 5)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rss"
                ],
                "summary": "Get top N SPIEGEL RSS headlines",
                "parameters": [
                    {
                        "maximum": 5,
                        "minimum": 1,
                        "type": "integer",
                        "default": 5,
                        "description": "Number of headlines to fetch (1-5)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter headlines by keyword",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HeadlinesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Unable to fetch RSS feed"
                }
            }
        },
        "handlers.GreetResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello, World!"
                }
            }
        },
        "handlers.HeadlinesResponse": {
            "type": "object",
            "properties": {
                "headlines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shared.RssHeadline"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
        "shared.RssHeadline": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "publishedAt": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}